No base erase

+++ entermediadb-deploy.sh
@@ -46,7 +46,14 @@ if [[ ! -d $WEBAPP/WEB-INF/data/system ]]; then
 fi
 
 ##Always replace the base and lib folders on new container
-rsync -ar --delete --exclude '/WEB-INF/data'  --exclude '/WEB-INF/encrypt.properties'  --exclude '/WEB-INF/pluginoverrides.xml' --exclude '/WEB-INF/classes' --exclude '/WEB-INF/elastic'  $EMCOMMON/webapp/WEB-INF $WEBAPP/
+
+if [[ ! -d $WEBAPP/WEB-INF/base ]]; then
+	rsync -ar --delete --exclude '/WEB-INF/data'  --exclude '/WEB-INF/encrypt.properties'  --exclude '/WEB-INF/pluginoverrides.xml' --exclude '/WEB-INF/classes' --exclude '/WEB-INF/elastic'  $EMCOMMON/webapp/WEB-INF $WEBAPP/
+fi
+
+##always upgrade
+rsync -ar --delete $EMCOMMON/webapp/WEB-INF/bin $WEBAPP/WEB-INF/
+rsync -a $EMCOMMON/webapp/WEB-INF/web.xml $WEBAPP/WEB-INF/web.xml
 
 
 if [[ ! -d $EMTARGET/tomcat/conf ]]; then
